Close #49 Change category implemented
