+++ TaskAppsAssasinsIOS/Controller/Note/NoteNibTableViewCell.swift
@@ -26,5 +26,9 @@ class NoteNibTableViewCell: UITableViewCell {
 
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)
+        
+        print(selected)
+        
+        
     }
 }
