+++ TaskAppsAssasinsIOS/Controller/Note/NoteViewController.swift
@@ -123,6 +123,11 @@ class NoteViewController: UIViewController {
         picturesEntity = loadImagesByNote()
         sortNameButton.layer.cornerRadius = 4
         sortDateButton.layer.cornerRadius = 4
+        
+        
+        
+        var longPressGes = UILongPressGestureRecognizer(target: self, action: #selector(invokeCategoryMenu))
+        noteTableView.addGestureRecognizer(longPressGes)
     }
 }
 
@@ -134,6 +139,7 @@ extension NoteViewController: UITableViewDelegate, UITableViewDataSource {
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = noteTableView.dequeueReusableCell(withIdentifier: "NoteNibTableViewCell", for: indexPath) as! NoteNibTableViewCell
+        cell.accessoryType = .disclosureIndicator
 
         cell.titleLabel?.text = filteredNotes[indexPath.row].title
         cell.descriptionLabel?.text = filteredNotes[indexPath.row].noteDescription
@@ -146,9 +152,40 @@ extension NoteViewController: UITableViewDelegate, UITableViewDataSource {
         return cell
     }
     
+    
+    @objc func invokeCategoryMenu(gesture: UIGestureRecognizer) {
+        if gesture.state == .began {
+            let touchPoint = gesture.location(in: noteTableView)
+            if let indexPath = noteTableView.indexPathForRow(at: touchPoint) {
+                print("LONG PRESS INSIDE CELL \(indexPath.row)")
+                
+                
+                // TODO: Implement Change Category
+                /*
+                let changeCategoryMenu = ChangeCategoryMenu()
+                changeCategoryMenu.noteViewControllerDelegate = self
+                changeCategoryMenu.categories = self.fetchAllCategory()
+                changeCategoryMenu.noteToChange = self.filteredNotes[indexPath.row]
+                */
+                //present(changeCategoryMenu, animated: true)
+                /*
+                let switchCategoryVC = ChangeCategoryView()
+                switchCategoryVC.noteViewControllerDelegate = self
+                switchCategoryVC.categories = self.fetchAllCategory()
+                switchCategoryVC.noteToChange = self.filteredNotes[indexPath.row]
+
+                switchCategoryVC.modalPresentationStyle = .overFullScreen
+                self.present(switchCategoryVC, animated: false)
+                 */
+            }
+        }
+    }
+    
+
+    
     func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
 
-        var noteCell = tableView.dequeueReusableCell(withIdentifier: "NoteNibTableViewCell", for: indexPath) as! NoteNibTableViewCell
+        //var noteCell = tableView.dequeueReusableCell(withIdentifier: "NoteNibTableViewCell", for: indexPath) as! NoteNibTableViewCell
         
         let deleteAction = UIContextualAction(style: .destructive, title: nil, handler: {(action, view, completionHandler) in
             let alertController = UIAlertController(title: "Delete", message: "Are you sure?", preferredStyle: .actionSheet)
@@ -179,14 +216,7 @@ extension NoteViewController: UITableViewDelegate, UITableViewDataSource {
         deleteAction.image = UIImage(systemName: "trash")
 
         let edit = UIContextualAction(style: .normal, title: "Edit", handler: {(action, view, completionHandler) in
-            // TODO: Implement Change Category
-            let switchCategoryVC = ChangeCategoryView()
-            switchCategoryVC.noteViewControllerDelegate = self
-            switchCategoryVC.categories = self.fetchAllCategory()
-            switchCategoryVC.noteToChange = self.filteredNotes[indexPath.row]
-
-            switchCategoryVC.modalPresentationStyle = .overFullScreen
-            self.present(switchCategoryVC, animated: false)
+            
             
             completionHandler(true)
         })
@@ -226,6 +256,25 @@ extension NoteViewController: UITableViewDelegate, UITableViewDataSource {
             
             self.navigationController?.pushViewController(noteDetailViewController, animated: true)
         }
+    }
+    
+    func tableView(_ tableView: UITableView, contextMenuConfigurationForRowAt indexPath: IndexPath, point: CGPoint) -> UIContextMenuConfiguration? {
+        
+        let changeCategoryMenu = ChangeCategoryMenu()
+        changeCategoryMenu.noteViewControllerDelegate = self
+        changeCategoryMenu.categories = self.fetchAllCategory()
+        changeCategoryMenu.noteToChange = self.filteredNotes[indexPath.row]
+        
+        let config = UIContextMenuConfiguration(identifier: nil, previewProvider: nil, actionProvider:  { _ -> UIMenu? in
+            
+            let menu = UIMenu(title: "Chance Category", image: UIImage(systemName: "pencil.circle") ,children: changeCategoryMenu.categoryOptions())
+            return menu
+        })
+        return config
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        noteTableView.reloadData()
     }
 }
